New version (1.99) to prepare the version 2.0.

+++ src/device/DataInfo.java
@@ -2,7 +2,7 @@ package device;
 
 public class DataInfo {
 
-	public static int UartDataRate = 9600;// 9600 bits/s
-	public static int ChDataRate = 250000; //250k bits/s 
+	public static long UartDataRate = 9600;// 9600 bits/s
+	public static long ChDataRate = 250000; //250k bits/s 
 	
 }
