+++ src/script/Command_ATCH.java
@@ -15,13 +15,13 @@ public class Command_ATCH extends Command {
 	}
 
 	@Override
-	public int execute() {
+	public long execute() {
 		SimLog.add("S" + sensor.getId() + " ATCH "+arg);
 		String args = sensor.getScript().getVariableValue(arg);
 		sensor.setCh(Integer.valueOf(args));
 		double ratio = (DataInfo.ChDataRate*1.0)/(DataInfo.UartDataRate);
 		String message = "ATCH "+args;
-		return (int)(Math.round(message.length()*8.*ratio));
+		return (long)(Math.round(message.length()*8.*ratio));
 	}
 	
 	@Override
